backend/requirements.txt
# CareCompanion FastAPI Backend
# Install: pip install -r requirements.txt

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.13.3

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
cryptography==43.0.3

# HTTP Client (wearables, HawkEye, UIDAI)
httpx==0.27.2

# SMS (Emergency contacts)
twilio==9.3.7

# ML (Phase 3 — uncomment when training/deploying models)
# scikit-learn==1.5.2
# xgboost==2.1.1
# numpy==2.1.3
# pandas==2.2.3
# joblib==1.4.2

# Dev / Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-multipart==0.0.12
